+++ R/utils.R
@@ -282,13 +282,19 @@ strcat <- function(a, b) .Call(rnng_strcat, a, b)
 #'     to be sent and received between different R sessions.
 #'
 #' @param refhook \strong{either} a list or pairlist of two functions: the
-#'     signature for the first must accept a list of reference objects (which
-#'     inherit 'class') and return a raw vector, e.g.
-#'     \code{torch::torch_serialize}, and the second must accept a raw vector
-#'     and return a list of reference objects, e.g. \code{torch::torch_load},
+#'     signature for the first must accept a reference object inheriting from
+#'     'class' (or a list of such objects) and return a raw vector, and the
+#'     second must accept a raw vector and return reference objects (or a list
+#'     of such objects),
 #'     \cr \strong{or else} NULL to reset.
-#' @param class [default 'torch_tensor'] a character string representing the
-#'     class of object that these serialization function will be applied to.
+#' @param class [default ''] a character string representing the
+#'     class of object that these serialization function will be applied to,
+#'     e.g. 'ArrowTabular' or 'torch_tensor'.
+#' @param list [default FALSE] the serialization functions accept and return
+#'     reference object individually e.g. \code{arrow::write_to_raw} and
+#'     \code{arrow::read_ipc_stream}. If TRUE, the serialization functions
+#'     accept and return a list of reference objects, e.g.
+#'     \code{torch::torch_serialize} and \code{torch::torch_load}.
 #' @param mark [default FALSE] (for advanced use only) logical value, whether to
 #'     mark serialized data with a special bit.
 #'
@@ -296,6 +302,8 @@ strcat <- function(a, b) .Call(rnng_strcat, a, b)
 #'
 #' @details Calling this function without any arguments returns the pairlist of
 #'     currently-registered 'refhook' functions (and resets 'mark' to FALSE).
+#'
+#'
 #'
 #' @examples
 #' g <- next_config(refhook = list(function(x) serialize(x, NULL), unserialize))
@@ -307,5 +315,5 @@ strcat <- function(a, b) .Call(rnng_strcat, a, b)
 #'
 #' @export
 #'
-next_config <- function(refhook = list(), class = "torch_tensor", mark = FALSE)
-  .Call(rnng_next_config, refhook, class, mark)
+next_config <- function(refhook = list(), class = "", list = FALSE, mark = FALSE)
+  .Call(rnng_next_config, refhook, class, list, mark)